Refactor code to move skill activation as counter attack parsing into domain model (#251).
