+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.spec.ts
@@ -213,66 +213,6 @@ describe('PassiveEffectParser', () => {
     });
   });
 
-  it('should parse counter attack with skill effect with max per turn', () => {
-    // GIVEN
-    const skills = JSON.parse(PASSIVE_SKILLS_TEST_DATA);
-    const skill: Skill = skills['100020'];
-    const names = JSON.parse(ABILITY_SKILLS_NAMES_TEST_DATA);
-    skill.names = names['100020'];
-    const descriptions = JSON.parse(ABILITY_SKILLS_SHORTDESCRIPTIONS_TEST_DATA);
-    skill.descriptions = descriptions['100020'];
-
-    const effect = JSON.parse('[0, 3, 49, [15, 3, 100020, 2]]');
-    const skillsServiceMock = new SkillsServiceMock() as SkillsService;
-    SkillsService['INSTANCE'] = skillsServiceMock;
-    spyOn(skillsServiceMock, 'searchForSkillByGumiId').and.returnValue(Skill.produce(skill));
-    // WHEN
-    const s = PassiveEffectParserFactory.getParser(effect[0], effect[1], effect[2]).parse(effect, null);
-    // THEN
-    expect(s).toEqual('15% de chance de contrer les dégâts physiques par: +20% PV (max 2 par tour)');
-  });
-
-  it('should parse counter attack with skill effect without max limit per turn', () => {
-    // GIVEN
-    const skills = JSON.parse(PASSIVE_SKILLS_TEST_DATA);
-    const skill: Skill = skills['100020'];
-    const names = JSON.parse(ABILITY_SKILLS_NAMES_TEST_DATA);
-    skill.names = names['100020'];
-    const descriptions = JSON.parse(ABILITY_SKILLS_SHORTDESCRIPTIONS_TEST_DATA);
-    skill.descriptions = descriptions['100020'];
-
-    const effect = JSON.parse('[0, 3, 50, [15, 3, 100020]]');
-    const skillsServiceMock = new SkillsServiceMock() as SkillsService;
-    SkillsService['INSTANCE'] = skillsServiceMock;
-    spyOn(skillsServiceMock, 'searchForSkillByGumiId').and.returnValue(Skill.produce(skill));
-    // WHEN
-    const s = PassiveEffectParserFactory.getParser(effect[0], effect[1], effect[2]).parse(effect, null);
-    // THEN
-    expect(s).toEqual('15% de chance de contrer les dégâts magiques par: +20% PV');
-  });
-
-  it('should parse counter attack with skill with chain damages', () => {
-    // GIVEN
-    const skills = JSON.parse(ABILITY_SKILLS_TEST_DATA);
-    const skill: Skill = skills['509624'];
-    const names = JSON.parse(ABILITY_SKILLS_NAMES_TEST_DATA);
-    skill.names = names['509624'];
-    const descriptions = JSON.parse(ABILITY_SKILLS_SHORTDESCRIPTIONS_TEST_DATA);
-    skill.descriptions = descriptions['509624'];
-    skill.active = true;
-    skill.gumi_id = 509624;
-
-    const effect = JSON.parse('[0, 3, 49, [15, 3, 509624]]');
-    const skillsServiceMock = new SkillsServiceMock() as SkillsService;
-    SkillsService['INSTANCE'] = skillsServiceMock;
-    spyOn(skillsServiceMock, 'searchForSkillByGumiId').and.returnValue(Skill.produce(skill));
-    // WHEN
-    const s = PassiveEffectParserFactory.getParser(effect[0], effect[1], effect[2]).parse(effect, null);
-    // THEN
-    expect(s).toEqual('15% de chance de contrer les dégâts physiques par: ' +
-      '<a href="ffexvius_skills.php?gumiid=509624">Lame des braves (FFV)</a>');
-  });
-
   it('should parse multi-skill when effect is the only one', () => {
     // GIVEN
     const skills = JSON.parse(ABILITY_SKILLS_TEST_DATA);
